Support export of QAT finalized models
Fix converting trainable quantizers to inferable quantizers in ActivationQuantizationHolder.
The issue with the current status is that finalized models can not be saved and loaded since
there are weights in the layer that no longer exist after converting to an inferable quantizer.
This commit fixes this issue by creating a new layer from the configuration of the layer
(after the trainable->inferable conversion).

Also, this commit fixes an issue with activation inferable quantizers that should receive a
list of thresholds and not a numpy array.